#24 모임 참여자를 확인할 수 있는 서비스 주체 변경
* 기존 guest 서비스의 API를 host 서비스로 이전
  * 모임 참여자를 볼 수 있는건 host 입장이기 때문
